+++ config/expert/Hourglass/code/CGCode.cpp
@@ -1,0 +1,21 @@
+#include <iostream>
+#include <string>
+#include <vector>
+#include <algorithm>
+
+using namespace std;
+
+int main()
+{
+    for (int i = 0; i < 23; i++) {
+        string h_line;
+        getline(cin, h_line);
+    }
+    int n;
+    cin >> n; cin.ignore();
+
+    // Write an answer using cout. DON'T FORGET THE "<< endl"
+    // To debug: cerr << "Debug messages..." << endl;
+
+    cout << "answer" << endl;
+}