+++ config/expert/Hourglass/code/CGCode.java
@@ -1,0 +1,19 @@
+import java.util.*;
+import java.io.*;
+import java.math.*;
+
+class Solution {
+
+    public static void main(String args[]) {
+        Scanner in = new Scanner(System.in);
+        for (int i = 0; i < 23; i++) {
+            String hLine = in.nextLine();
+        }
+        int N = in.nextInt();
+
+        // Write an answer using System.out.println()
+        // To debug: System.err.println("Debug messages...");
+
+        System.out.println("answer");
+    }
+}