+++ config/expert/Hourglass/code/CGCode.js
@@ -1,0 +1,9 @@
+for (let i = 0; i < 23; i++) {
+    const hLine = readline();
+}
+const N = parseInt(readline());
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');