+++ config/expert/Hourglass/code/CGCode.ts
@@ -1,0 +1,9 @@
+for (let i = 0; i < 23; i++) {
+    const hLine: string = readline();
+}
+const N: number = parseInt(readline());
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');